bilbo and Mummi contribution!! veriverse contribution!! https://LiveEdu.tv/ (Livecoding.tv) contribution!! ThomasBS (Thomas Boregaard Sørensen, https://kaem.dk/, https://kaem.io/, tbhs1000%💛forgrowing!! Den bedste udvikler!), drmjg, camwhite (Cameron White, Twitchs' scamwhite), xmetrix, intelijens, syaseigirl, CatherinaHelena (CatherinaCortega, Twitchs' helenawinter), fuel, r00tkiller, hiiamgosu (itashellhaha...(gosu), scorpioinyouranus... ), pablosj (sjpablo, Pawel Kulesza, Pablosj special contribution (after so much days with so low audience... ) .. contribution!!), phanxgames, 0din, anita, beater, benjones, bill_gates, brayoeh, camilosasuke (me :ca2:), cary, clementin, Cookiesliyr, costinis, crisy, dauntlesstech, david_joffe, denvercowboy, drcode, dslegends, ederangelo21..., epryka, evgeniy_andreev, famed, flythiago, frankly444, georgeb, gigaherz, gostrafx, guacjack, hatyme, heinz3, hosjenken, intelijens, irwanbd, isabella, jackperrie, jacobs, jaycee, kaeris, kannu, karrmarr, kayne3, kissloryshy, klerind, kozmo, kranalex, kruzoe, landry96, larrybrook, LiveEdu_bot, liveedu_support, LiveOrDevTrying, lyndroid, maoabc1818, mark_miller, masterchoc, matroskinb, mitch_ill, mitsurugi, nero951, ngetich, novicehere, olivert, olelukoe, pestka12, pinkgorilla, pirateofthesea, polodonald, rex2, resident, riverside, rooney2, samdev, sargin, shgalym (from Kazakhstan), shieldgenerator7 (Twitchtv shieldgenerator7), shinkaisokuuindo, siox, starlord91, stefan_v, stephen22, systemis, thiagovgamg, tonycode, valdinei, vapaamies, vodkavodkavodka228, wariosolis, wetriks, wilson2000, yanaseattle, zakaria98, zzeroge and zzzel contribution!! https://Twitch.tv/ contribution!! ! ThomasBS_ (LiveEdu.tv ThomasBS, Den besten Software udvikler), RocketBeans.TV (Nils, Simon, Colin, Donni, Etienne, Florentin, Lars, Marco, Oliver, Renè, [Schröck, Daniel] und Viet), AlfredoPlays (and his mates... Alfredo, Hanson, Rey and Mackie contribution), KindaFunnyGames (Greg Miller, Nick Scarpino, Andy and Kevin contribution...), SoulFoamTV (LettuceSmokeStudios, SoulFoam), scamwhite (Cameron White, Liveedutvs' camwhite, mesh.io), Phanxgames, Peter_Alfons_Loch (from phanxgames streamd and discord channel, first "explicit" and "spontaneous" commit and contribution to this newly create GitHub repository), cancerMoonLife(hiiamgosu from LiveEdu), Spoofh (LiveEdu.tv Fuel: Benjamin Kuhl), themushmouth (LiveEdu.tv mushmouth: Patrick McFarling), Tectuma (Kevin, Stacie (GothicCat1), Matt Brock (Cronos) and Dylan and Nick...), IILomoo (Liveedu lomoo, Matti), nothings2 (Sean Barret), MitchJones, sonnystar (hairy green sonny star, Vlad Ionescu), helenawinter (CatherinaHelena, CatherinaCortega from LiveEdu), XanderGamingTwich and xmxstudio (LiveCoding/LiveEdu xmetrix), Adan_204, Adevel, Always_Betonentei, Cylverrose, david_joffe, emielrregis, faillehad (french swearer with igrectepe) ), foldm, GameDuells, Grononmc, guy_ross, Hazurl, igrectepe (French swearer with faillehad) , itechno2016 (Russian guy), Javo_net, knutaf, Lauheid1974, Lessio__, LiveOrDevTrying, LoryMaster, OrcaPrime, OxenHill, Piepassaro, Pixel701, Pucki4, recursionishard, shieldgenerator7 (Liveedutv shieldgenerator7), SystemOfPeter, Tanerax, theformemrcc, Tookietookie, Trump88, Volon_, Vulsen, WildN00b, Wraithone,  contribution!! https://mixer.com/ (previous Beam.pro) contribution!! R4ver contribution (and his friends agile as him, from Sverige... contribution), ArrestedMoney2, LlamaJuic3467, NormalZeus53319, sharkshooter641, TaxingBinkie154 and WedRhombus23 contribution!! YouTube Live contribution!! Shivam Bhola contribution!! bisba contribution!! Carol, Lúcio and Lucinho contribution!! Upwork.com contribution!! Bock Engineering contribution!!
Flight Gauge Prober - Projects On Demand
from bilbo at veriverse!! Internet Net Curitiba 60Mbps▼ 6Mbps▴
i7 6700 Dell XPS 8900 32GB DDR4 RAM 2x240GB SSD 2TB HDD
Windows 10 Pro (Quad Boot with Windows 10 Pro with Hyper-V, Windows 8.1 Pro, macOS Sierra 10.12.6 and Ubuntu 17.04)

+++ appseed/aura/primitive/math/math_department.h
@@ -6,62 +6,62 @@ namespace math
 
 
    class CLASS_DECL_AURA math :
-   virtual public ::aura::department
+      virtual public ::aura::department
    {
-   private:
+      private:
 
 
-      double         dPi;
-      //#if WINVER >= 0x500
+         double         dPi;
+         //#if WINVER >= 0x500
 #ifdef WINDOWSEX
-      HCRYPTPROV     hCryptProv;
-      HCRYPTKEY      hOriginalKey;
-      HCRYPTKEY      hDuplicateKey;
+         HCRYPTPROV     hCryptProv;
+         HCRYPTKEY      hOriginalKey;
+         HCRYPTKEY      hDuplicateKey;
 #endif
-      //#endif
-      uint32_t       dwMode;
-      BYTE           pbData[16];
-      rng            m_rng;
+         //#endif
+         uint32_t       dwMode;
+         BYTE           pbData[16];
+         rng            m_rng;
 
 
-   public:
+      public:
 
 
-      mutex          m_mutex;
+         mutex          m_mutex;
 
 
-      math(::aura::application * papp);
-      virtual ~math();
+         math(::aura::application * papp);
+         virtual ~math();
 
 
-      int64_t RandRange(int64_t iMin, int64_t iMax);
-      double frand_range(double iMin, double iMax);
+         int64_t RandRange(int64_t iMin, int64_t iMax);
+         double frand_range(double iMin, double iMax);
 
 
-      int32_t LinearMap(int32_t iMin, int32_t iMax, int32_t iValue, int32_t iValueMin, int32_t iValueMax);
-      uint64_t gen_rand();
-      void random_bytes(void * buf, memory_size_t dwLen);
-      void RandomBytes(void * buf, memory_size_t dwLen);
-      void gen_rand(void * buf, memory_size_t dwLen);
-      void gen_rand_alnum(char * buf, memory_size_t dwLen);
-      uint32_t rnd();
-      int32_t random_context_entropy(int32_t iMin, int32_t iMax, int32_t iLevel = 3);
-      int32_t rand_max();
-      double LinearMap(
-         double dMin, double dMax,
-         double dValue,
-         double dValueMin,
-         double dValueMax);
-      double GetPi();
+         int32_t LinearMap(int32_t iMin, int32_t iMax, int32_t iValue, int32_t iValueMin, int32_t iValueMax);
+         uint64_t gen_rand();
+         void random_bytes(void * buf, memory_size_t dwLen);
+         void RandomBytes(void * buf, memory_size_t dwLen);
+         void gen_rand(void * buf, memory_size_t dwLen);
+         void gen_rand_alnum(char * buf, memory_size_t dwLen);
+         uint32_t rnd();
+         int32_t random_context_entropy(int32_t iMin, int32_t iMax, int32_t iLevel = 3);
+         int32_t rand_max();
+         double LinearMap(
+            double dMin, double dMax,
+            double dValue,
+            double dValueMin,
+            double dValueMax);
+         double GetPi();
 
-      static bool IsPowerOfTwo(uint32_t uiValue);
-      static uint32_t ReverseBits(uint32_t index, uint32_t NumBits);
-      //   template <class N> static inline void MaxClip(N * pnClipValue, N nMaxValue);
-      //   template <class N> static inline void MinClip(N * pnClipValue, N nMinValue);
-      static inline void MaxClip(double * pnClipValue, double nMaxValue);
-      static inline void MinClip(double * pnClipValue, double nMinValue);
+         static bool IsPowerOfTwo(uint64_t uiValue);
+         static uint32_t ReverseBits(uint32_t index, uint32_t NumBits);
+         //   template <class N> static inline void MaxClip(N * pnClipValue, N nMaxValue);
+         //   template <class N> static inline void MinClip(N * pnClipValue, N nMinValue);
+         static inline void MaxClip(double * pnClipValue, double nMaxValue);
+         static inline void MinClip(double * pnClipValue, double nMinValue);
 
-      int32_t RangeRate(int32_t iMin, int32_t iMax, double dRate);
+         int32_t RangeRate(int32_t iMin, int32_t iMax, double dRate);
